Update chekout.php
1) For normal or custom billing and shipping fields you can use woocommerce_billing_fields or woocommerce_shipping_fields action hooks on checkout page as follow.

It will make a custom checkout field required without any need to add a validation script and to save it in the order. The field will also appear in My account edit adresses field section.

Some argument explanations:

The class 'update_totals_on_change' allow to trigger "update checkout" on change.

The 'required' attribute make the field required or not

The 'priority' attribute allow you to change the location of the field.